requirements.txt
# Core Framework
streamlit==1.31.0
python-dotenv==1.0.1

# LangChain and LangSmith
langchain==0.1.9
langchain-community==0.0.24
langsmith==0.1.0

# LLM and Embeddings
ollama==0.1.7
sentence-transformers==2.3.1

# Vector Database
chromadb==0.4.22

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
pandas==2.2.0

# Web Scraping
beautifulsoup4==4.12.3
lxml==5.1.0
requests==2.31.0
selenium==4.17.2

# Database
sqlalchemy==2.0.25

# Export
reportlab==4.0.9
fpdf2==2.7.7

# Image Processing
Pillow==10.2.0
pytesseract==0.3.10
easyocr==1.7.1

# Cloud Providers
google-generativeai==0.3.2

# Utilities
validators==0.22.0
python-magic==0.4.27
urllib3==2.2.0
